Feat/proper grpc init (#89)
* feat(boot): redis proper shutdown

* feat(grpc-boot): register grpc controllers in main

* feat(grpc-boot): tollgate controller added

* feat(grpc-boot): tollgate crossing grpc minor renaming

* feat(grpc-boot): review feedback addressing

+++ internal/service/tollgate/detector/bbox_test.go
@@ -31,7 +31,7 @@ func TestDetectCrossingBBox(t *testing.T) {
 }
 
 func testCrossing(ctx context.Context, t *testing.T, s *storage) {
-	required := 4
+	required := int32(4)
 	bBoxes := []*BBox{{
 		LonMin: 1,
 		LatMin: 1,
